Carrito local completo, cambios a actions y redux y UI de usuario no logeado

+++ Client/src/Redux/Actions/Actions.js
@@ -23,7 +23,7 @@ export const SHOW_PRODUCTS = "SHOW_PRODUCTS";
 export const getProducts = () => {
     return async (dispatch) => {
         const response = await axios.get("/products");
-        console.log("response: ", response);
+        const products = response.data?.filter(p => p.status !== false)
         return dispatch({
             type: GET_PRODUCTS,
             payload: response.data,
@@ -64,10 +64,8 @@ export const getProductById = (id) => {
 };
 
 export const getCategories = () => {
-    console.log("pasé actions")
     return async (dispatch) => {
         const response = await axios.get("/categorys");
-        console.log("response: ", response);
         return dispatch({
             type: GET_CATEGORIES,
             payload: response.data,
@@ -92,25 +90,46 @@ export const getProductsByMinMax = (min, max) => {
 
 // add product to cart, if the product is already in the cart, increase the quantity by +1
 export const addProduct = (data) => {
-    console.log("Desde la action: ", data)
     return async (dispatch) => {
-        if (data.user_id) {
+        if(data.user_id) {
+            console.log("Estoy agregando productos: ", data)
             await axios.post("/shoppingCarts/addProductToShoppingCart", data)
-                .then((response) => {
+            .then((response) => {
                     console.log(response.data)
                     dispatch({ type: "ADD_PRODUCT", payload: response.data })
                 })
-        };
-    }
+            } else {
+                const product_id = data.products_id[0].product_id
+                const product = await axios.get(`/products//detail/${product_id}`)
+                const localCart = {
+                    user_id: "LocalCart",
+                    products: [{
+                        product_id: product.data,
+                        quantity: data.products_id[0].quantity
+                    }]
+                }
+            dispatch({
+                type: "ADDPRODUCT_LOCALCART",
+                payload: localCart
+            })
+        }
+    };
 }
+
 // get cart by user
 export const getCart = (id) => {
     return async (dispatch) => {
-        await axios.get(`/shoppingCarts/detail/${id}`)
-            .then((response) => {
-                console.log("response.data: ", response.data)
-                dispatch({ type: "GET_CART", payload: response.data })
+        if(id) {
+            await axios.get(`/shoppingCarts/detail/${id}`)
+                .then((response) => {
+                    console.log("response.data: ", response.data)
+                    dispatch({ type: "GET_CART", payload: response.data })
+                })
+        } else {
+            dispatch({
+                type: "GET_CART"
             })
+        }
     };
 }
 
@@ -118,11 +137,22 @@ export const getCart = (id) => {
 export const removeQuantity = (data) => {
     console.log("Desde la action: ", data)
     return async (dispatch) => {
-        await axios.post("/shoppingCarts/deleteProductFromShoppingCart", data)
-            .then((response) => {
-                console.log(response.data)
-                dispatch({ type: "REMOVE_QUANTITY", payload: response.data })
+        if(data.user_id) {
+            await axios.post("/shoppingCarts/deleteProductFromShoppingCart", data)
+                .then((response) => {
+                    console.log(response.data)
+                    dispatch({ type: "REMOVE_QUANTITY", payload: response.data })
+                })
+        } else {
+            const localCart = {
+                user_id: "LocalCart",
+                product_id: data.product_id,
+            }
+            dispatch({
+                type: "REMOVEQUANTITY_LOCALCART",
+                payload: localCart
             })
+        }
     };
 }
 
@@ -130,24 +160,37 @@ export const removeQuantity = (data) => {
 export const removeProduct = (data) => {
     console.log("desde la action: ", data)
     return async (dispatch) => {
-        await axios.post("/shoppingCarts/deleteProductFromShoppingCartAndDeleteShoppingCart", data)
-            .then((response) => {
-                console.log(response.data)
-                dispatch({ type: "REMOVE_PRODUCT", payload: response.data })
+        if(data.user_id) {
+            await axios.post("/shoppingCarts/deleteProductFromShoppingCartAndDeleteShoppingCart", data)
+                .then((response) => {
+                    console.log(response.data)
+                    dispatch({ type: "REMOVE_PRODUCT", payload: response.data })
+                })
+        } else {
+            const localCart = {
+                user_id: "LocalCart",
+                products: data.product_id,
+            }
+            dispatch({
+                type: "REMOVEPRODUCT_LOCALCART",
+                payload: localCart
             })
+        }
     };
 }
 
 // delete cart
 export const removeCart = (id) => {
     return async (dispatch) => {
-        await axios.delete(`/shoppingCarts/deleteShoppingCart/${id}`)
-            .then((response) => {
-                dispatch({ type: "REMOVE_CART", payload: [] })
-            })
-        let quantity = { quantity: 0 }
-        dispatch({ type: "GET_QUANTITY", payload: quantity })
-
+        if(id) {
+            await axios.delete(`/shoppingCarts/deleteShoppingCart/${id}`)
+                .then((response) => {
+                    dispatch({ type: "REMOVE_CART", payload: [] })
+                })
+                let quantity = { quantity: 0 }
+                dispatch({ type: "GET_QUANTITY", payload: quantity })
+        }
+        dispatch({ type: "REMOVE_CART", payload: [] })
     };
 }
 
@@ -155,10 +198,14 @@ export const removeCart = (id) => {
 
 export const getQuantity = (id) => {
     return async (dispatch) => {
-        await axios.get(`/shoppingCarts/quantity/${id}`)
+        if(id) {
+            await axios.get(`/shoppingCarts/quantity/${id}`)
             .then((response) => {
                 dispatch({ type: "GET_QUANTITY", payload: response.data })
-            })
+                })
+        } else {
+            dispatch({ type: "GET_QUANTITY" })
+        }
     };
 }
 
@@ -167,20 +214,24 @@ export const getQuantity = (id) => {
 export const getFavorites = (id) => {
     // let id = "63681baa20ab92251bb85fd9"
     return async (dispatch) => {
-        await axios.get(`/users/favorites/${id}`)
-            .then((response) => {
-                dispatch({ type: "GET_FAVORITES", payload: response.data })
-            })
+        if(id) {
+            await axios.get(`/users/favorites/${id}`)
+                .then((response) => {
+                    dispatch({ type: "GET_FAVORITES", payload: response.data })
+                })
+        }
     };
 }
 
 // add favorite
 export const addFavorite = (data) => {
     return async (dispatch) => {
-        await axios.post("/users/favorites", data)
-            .then((response) => {
-                dispatch({ type: "ADD_FAVORITE", payload: response.data })
-            })
+        if(data) {
+            await axios.post("/users/favorites", data)
+                .then((response) => {
+                    dispatch({ type: "ADD_FAVORITE", payload: response.data })
+                })
+        }
     };
 }
